+++ UserBookSubscribeAPI/Models/DTO/SubscribeDTO.cs
@@ -3,7 +3,7 @@ namespace UserBookSubscribeAPI.Models.DTO
 {
     public class SubscribeDTO
     {
-        public int UserID { get; set; }
-        public int BookID { get; set; }
+        public long UserID { get; set; }
+        public long BookID { get; set; }
     }
 }
